+++ corpus/cases/adv-foreign-namespace/utils.py
@@ -1,0 +1,2 @@
+def new_helper():
+    return 1