+++ corpus/cases/adv-shadowed-and-template/pylint/plugins/myplug/checker.py
@@ -1,0 +1,5 @@
+from pylint.checkers import BaseChecker
+
+
+class MyChecker(BaseChecker):
+    pass